Need to finish the cleaning of incoming data
